Reduce token bucket rate limiter spinlock contention
An earlier version of this patch used a single spin lock to protect a
single token bucket.  A test with 20 threads putting 1000 byte values
resulted in each thread spending 65% of its cpu time in the spin lock.
With 96 threads: 88%.

This version uses an array of 7 token bucket rate limiters, each with
its own spin lock.  The configured burst and rate is divided equally
among them. Threads randomly choose a token bucket when requesting
tokens.

With this change, the same test never exceeds 2% cpu time in
the spinlock -- even with up to 1024 threads.