Fix misuse of void* in c0 ingest worker struct
It was set to 'struct c0sk_impl *' but used as if it was
a 'struct c0sk *' (i.e., a "handle" to the "impl").  Luckily
the "handle" is the first member in 'struct c0sk_impl' so no harm was
done.  If the interfaces and struct members were typed, this would've
been caught by the compiler.

A great lesson in the dangers of 'void *' -- the compiler can't help you.