Implement HSE runtime home
HSE runtime home is a directory for the socket file, log file, and
global configuration file. The global configuration file called hse.conf
is different than what used to be hse.conf. The old hse.conf is now
called a kvdb.conf.

HSE runtime home can be the directory as a KVDB home. In fact, this is
probaby going to be the way most people use HSE. This includes all the
tests. hse.pid got renamed to kvdb.pid to match the kvdb-ness of KVDB
home.