Remove root mdc in favor of JSON metadata file
The JSON metadata file tracks the same data as the root MDC, but in a
more user-friendly way. In addition, the staging path and capacity path
are now being tracked as well, and will be used to determine staging and
capacity paths if they are not passed in as parameters.