Enable logging by default and fix crash
Crash was happening because of how enums currently work in this
param/config setup. String was a default, but the struct uses an enum
(int) to represent its data. Ideally all enums would use an enum (int)
in their struct definitions, but a day's worth of work would probably be
required to fix the params that use a string for an enum in their struct
definitions.