Add notion of KVDB alias
REST APIs typically are of the form of /resource/<specific resource>. In
HSE-land, that means we want something along the lines of /kvdb/<kvdb>.
As planned, the only unique identifier for a KVDB is its home
directory. Home directories are not candidates for being the identifier
in a REST path because it will have path separators. This also applies to
perf counters.

In order to mitigate this issue, an alias has been created. The alias is
an always incrementing integer. Every KVDB that is opened gets the
current alias value and then increments that value.