NFSE-5011: Add introspection APIs (#88)
* Define hse_static as HSE_WEAK in mocked builds

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add hse_kvdb_home_get()

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add hse_kvs_name_get()

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add stringification to parameters

This is in preparation for introspection APIs, but may prove more useful
in other contexts.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add infrastructure getting a parameter generically

All our various parameter types are essentially a thin wrapper over
generic infrastructure. Adding param_get() expands that infrastructure
in order to reduce the amount of work.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add hse_param_get()

Allows a user to introspect the stringified value of an hse global parameter.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add hse_kvdb_param_get()

Allows a user to introspect KVDB parameters.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add hse_kvs_param_get()

Allows a user to introspect KVS rparams and cparams.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add GET /kvdb/{alias}/home

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add GET /kvdb/{alias}/params & /kvdb/{alias}/params/{param}

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add GET /kvdb/{alias}/kvs/{name}/params & /kvdb/{alias}/kvs/{name}/params/{param}

Allows a user to query parameters of a KVS.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add jsonification of parameters

This allows HSE to turn a parameter struct back into its JSON
representation. One could have achieved this by cJSON_Parse()ing the
stringified version of the parameter, but that seemed needless in my
opinion. This makes the config code quite a bit more flexible.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add GET /params and /params/{param}

Allows for a user to query all parameters about HSE.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Warn about potential for undefined behavior in KVDB/KVS REST route registration

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Return false in default case for failures in release builds

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add generic infrastructure for setting a parameter generically

This will support setting writable parameters at runtime.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add PUT /params/{param}

Allows a user to change the value of a writable parameter in HSE
gparams.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add PUT /kvdb/{alias}/params/{param}

Allows a user a to modify writable KVDB rparams.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add PUT /kvdb/{alias}/kvs/{name}/params/{param}

Allows a user to modify writable KVS rparams.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Remove mapi_idx.h include from ikvdb_test.c

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Move media class name constants to public API

This is necessary for exposing the in future APIs.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Remove kvdb_rparams_resolve()

This was a holdover from a previous iteration of kvdb_rparams.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Use an abort() instead of assert(false)

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Format various parameter test

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add APIs to introspect media class information

The storage info APIs have been deprecated in favor of these new media
class info APIs. There has been a reduction in the amount of information
returned users because given a path, much of what the storage info APIs
provided could be discovered by users themselves using things statvfs()
et. al. Now we just return information that HSE itself could only know.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add documentation for REST interface

This document is using the OpenAPI specification. It can be viewed in
various editors on the web or in some cases editors with appropriate
extensions.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Fix logging_destination_converter() decl/def

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Fix comment for HSE_NONNULL

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Use abort() in param.c

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Use abort instead of assert(false) in config/logging.c

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Fix for loop parameters to use media class constants

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Remove unused code from mpool_test

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Review followup

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Add hse_mclass enum

Using strings to interact with the public API was too cumbersome.

Signed-off-by: Tristan Partin <tpartin@micron.com>

* Remove MP_MED* in favor of HSE_MCLASS*

Signed-off-by: Tristan Partin <tpartin@micron.com>