Add -Dbindings=auto and set it to the default
This is a step in right direction for getting TTC (time to compile) even
lower for external users/contributors. By making -Dbindings=auto the
default, we can check to make sure various requirements are met before
trying to compile various bindings.

In this case, we are checking if the version of Cython is >= 0.29.21
which is the documented version that hse-python supports.

In the case of Java, it would be: Is there a version of Java >= 8 that
we can use to compile and is Maven available.

In the case of Go, it would be: Is there a Go compiler available.

etc.

Signed-off-by: Tristan Partin <tpartin@micron.com>