Remove workqueue from REST subsystem
HTTP can be multiplexed depending on the number of connections. The
workqueue gets in the way of passing information back to the client. The
workqueue in this case was causing file descriptors to get closed and
requests cleaned up before HSE can even interact with the request. Now
requests are blocking, so resources stay active for as long as HSE
needs to process the request.

This commit is similar to some work that I had to do to actually test
writable parameters however long ago that was. No idea why I never
committed the work. Maybe it was because it felt like a large-ish change
and no one ever commented on it.

This commit specifically is really a stop-gap due to leaving the work
struct and thread args struct around, but I really don't want to spend
too much time in this code because I feel like I will rewrite it here
in the near future.

Signed-off-by: Tristan Partin <tpartin@micron.com>