Add an hblock to every kvset
An hblock stores various metadata about a kvset. A kvset can only ever
have one hblock.

Most of the data within the hblock is new in order to facilitate the
upcoming kvset split, but in some cases, data was moved from the last
kblock into the hblock.

In the past, the last kblock of a kvset stored information like the min
and max sequence numbers a kvset referenced, the prefix tombstone tree,
and the composite kvset hlog. This information now exists in the hblock.

Signed-off-by: Tristan Partin <tpartin@micron.com>