Fix the logic when setting up the ptree kblk_reader
This part of the code since the hblock was conceived has essentially
been one big hack for better or for worse. There has been a bit of
fallout in this area of the code. See the last two PRs by me.

In this case, we have a few changes that'll hopefully fix the stalling
being seen by Greg during mongo replication testing. `struct
kblk_reader` is a misnomer for the most part. Really it is a reader that
can span multiple WBTrees. In the case of the ptomb kblk_reader, there
is only ever one block to read. The ptomb WBTree is stored entirely in
the kvset's singular hblock. The bounds of kblk_reader::kr_kblk_cnt were
wrong in the case of the ptomb reader.

Signed-off-by: Tristan Partin <tpartin@micron.com>