Finish split key determination
Once we have found the kblock where we pass the 50% mark, we take all
the reverse kblock iterator offsets to create new forward leaf
iterators, where we march forward until we pass the 50% mark again. When
that point is reached, return the first key of that leaf node to the
caller.

Signed-off-by: Tristan Partin <tpartin@micron.com>