Reimplement REST support using libevent
Previously HSE used to use libmicrohttpd for its REST server. For
various reason that may or may not have been accurate, it was determined
that libmicrohttpd was not a good fit for what we were tring to
accomplish with a REST server in HSE.

Enter libevent. libevent is primarily an event loop library. It just so
happens that libevent also has support for an HTTP server. libevent is a
little bit more lighterweight than libmicrohttpd when it comes to
dependencies. libevent is also more useful than just an HTTP server. At
some point in time, we may want to integrate an event loop within HSE.
Who knows? But it is a possibility, and we wouldn't even need another
dependency for one.

I evaluated an alternative like libuv, which I happed to think might be
more maintained due to its use in technologies like node.js, but libuv
doesn't have an HTTP server, and instead only implements TCP. While the
HTTP 1.1 spec is fairly minimal, I don't have the time to implement it,
nor do I want other people to have to read it to contribute in this area
of the code.

I have created a libhse-rest.a for use within the rest of HSE. This
library provides the interfaces to register HTTP routes, create HTTP
request handlers, and lifecycle functions for an HTTP server. Many of
the interfaces are inspired by Go's HTTP support. libevent's structures
and APIs are hidden away from the users of the library should we ever
come to the determination that libevent no longer suits our needs. It
would be a fairly easy swap out of code contained with libhse-rest.a.

The hse_util/rest_client files have also been removed. These files made
HSE take a dependency on libcurl when the interfaces in
hse_util/rest_client weren't even used within the library. They were
used only in the CLI and maybe some tools. In order to replace the
functionality that was removed, I added a REST client into libhse-cli.a
in addition to REST API wrappers for C. The REST client is modeled a
little bit off of JavaScript's fetch() API.

The previous REST server was extremely under-documented. Routes,
methods, and parameters were fairly unknown. I believe you could query a
running REST server for some help documentation, but it just wasn't
cutting it. With this commit, I have added a docs/openapi.yaml file that
is a full description (or at least aims to be) of what the REST server
does and how to interact with it. OpenAPI is a fairly popular standard
for documenting HTTP APIs, so it was a fairly easy choice. In addition,
also added a GitHub Actions workflow for validating it in the event it
ever changes.

Many of the REST interfaces along with being re-implemented were cleaned
up. For instance, many used YAML. They all use JSON now. We set the
"Content-Type" header as appropriate. We return status codes which
describe the operations and their potential errors. This was previously
impossible in the previous iteration of the REST server. Some of the
output format changed, but all REST output has been thoroughly
documented with the OpenAPI description I have provided.

Signed-off-by: Tristan Partin <tpartin@micron.com>