Adjust logic related to n_committed in cn_mblocks_destroy()
It is kind of hard to tell if this logic was broken or not. In essence,
what we want to do, is delete all mblocks that have previously been
committed. This means we need to update n_committed every time we delete
an mblock, we need to decrement n_committed and track that for the
entire array of kvset_mblocks.

Signed-off-by: Tristan Partin <tpartin@micron.com>