Gracefully fallback when cmake does not exist
On platforms like Ubuntu 18.04, the version of CMake is too old for
Meson to use, so we can't actually build some of the tools that require
CMake-based subprojects. Unfortunately the logic was broken for a while,
but this should fix it.

Now when the cmake executable doesn't exist, we will gracefully degrade
the build instead of erroring out in the situation when -Dtools=auto.

Signed-off-by: Tristan Partin <tpartin@micron.com>