Use arch builds as an excuse to use subprojects
As we all know, creating cross compilation environments is not easy for
C and C++ projects. Recently we saw an issue that necessitated using
libcurl from the subproject instead of using it from the Debian repos
due to a package collision between libcurl4 and libcurl4:s390x.

This PR would force us to use subprojects on every alternative
architecture build and removes the need for the subprojects job.

Signed-off-by: Tristan Partin <tpartin@micron.com>