Fix the split key algorithm
Prior to this commit, this algorithm would find a split key that didn't
exist with 45%-55% range we were shooting for. The issue being that we
weren't seeking kblocks and WBT leaf nodes appropriately after finding
the inflection kblock.

In order to fix this, two new parts of the algorithm were added:

1. After finding the inflection kblock, move the reverse kblock
   iterators forward (not reverse) in order to find the first kblock
   for each iterator with a max key >= the min key of the inflection
   kblock.
2. Use the inflection key (inflection kblock's min key) as a comparator
   to find the starting WBT leaf node, where previously this index
   always started at 0.

Without both of these things, we ended up counting data we had never
actually seen during the iterations of the iterators.

In addition, the reverse kblock iterators used to decrement an extra
time instead of stopping where they were.

Signed-off-by: Tristan Partin <tpartin@micron.com>