Remove rarely used KVDB stats (NFSE-5303)
HSE maintains a small set of KVDB stats that are only used to show stats about
each KVS when the KVS is opened.  These KVDB stats are not maintained at
runtime, so they're not useful for anything except the log message during KVS
open.

The inaccurate KVDB stats have been removed, and the stats needed for the log
message are computed locally in cn_open().

Signed-off-by: Alex Tomlinson <4172734+alexttx@users.noreply.github.com>