Really fix the split key algorithm (#374)
Say the inflection key exists within the last WBT leaf node. We were
originally comparing the inflection key and the first key of the WBT
leaf node.

This is *extremely* wrong because we know that the inflection key (or a
key close to it) exists within the kblock, but by comparing the
inflection key to the first key, we didn't seek the forward WBT iterator
correctly. We *need* to find the first WBT leaf node whose largest key
(last key) is >= to the inflection key.

Signed-off-by: Tristan Partin <tpartin@micron.com>