Fix split key algorithm for realzies
Another huge problem that was theorized by Gaurav. The first part of the
algorithm includes finding the inflection points of each reverse kblock
iterator to take as input into the forward WBT leaf node iterators. The
issue was an incorrect understanding of what the inflection points
actually proved.

The inflection points prove that everything to the left of the points is
data that we can prove all kblocks at the end of the iterator are
greater than (keyspace-wise). So we accumulate all data which exists to
the left of every kblock within each kvset, and set that to the out
parameter seen_kvlen. This seen_kvlen is then provided as input into the
determination of the split key. Since we now know that all data to the
left of every kblock is accounted for, we can start the WBT leaf node
iterators at leaf index 0, and remove the WBT leaf node seek.

Signed-off-by: Tristan Partin <tpartin@micron.com>
Co-authored-by: Gaurav Ramdasi <gramdasi@micron.com>