Fix a few -Wconversion warnings (#430)
Wanted to play around with it, and we have a lot of issues. Some of them
are pretty bad too. uint -> int -> uint type stuff. Crazy!

Obviously it is hard to throw -Wconversion on all at once, so here is an
attempt at fixing a few issues that were pretty easy to reason about.

Signed-off-by: Tristan Partin <tpartin@micron.com>

Signed-off-by: Tristan Partin <tpartin@micron.com>