Make accessing data tree elements thread safe
`dt_find()` is used to find a data tree element based on the path.
Outside of unit tests, there is one single user, perfc. Perfc grabs the
dte and updates the user data of the dte in `perfc_alloc_impl()`. In
some theoretical land, there is a race condition in that the data tree
element may be deleted while perfc is updating it. `dt_update()` allows
for the element's user data to be updated while holding the tree lock.

Signed-off-by: Tristan Partin <tpartin@micron.com>