Fix ks->ks_kvsetid != 0 assert in kvset_open2()
The fix is to reset the hblock ID to zero in cn_split() after
calling delete_mblock() on it.

We used to rely on delete_mblock() to reset the blkid for us when
it was taking in a ptr to struct kvs_block. With the recent code
refactoring, delete_mblock() now takes in an mblock ID which puts
the onus on the caller to reset the blkid after delete_mblock().
To be safe, I've fixed all callers of delete_mblock().

Fix "modes" smoke test to run the EROFS/EPERM test only when
smoke is run as a non-root user.

Signed-off-by: Nabeel M Mohamed <50154757+nabeelmmd@users.noreply.github.com>