Fix a crash consistency bug in MDC open
A crash during MDC erase could leave the new passive log header partially
updated. On MDC reopen, we detect this crash situation correctly but the
passive log handle is erroneously freed causing MDC open to fail with
EINVAL.

The fix is to not free the passive log handle on detecting this crash
situation so that the subsequent erase of the passive log and the mdc
open succeeds.

Signed-off-by: Nabeel M Mohamed <50154757+nabeelmmd@users.noreply.github.com>
Reported-by: Yile (Michael) Gu