Initialize spn nodes at cn_tree_node creation. (#536)
Check and handle joined nodes that are on both the active and stable dirty lists.

This patch ensures spn nodes are always initialized, and now uses the "spn_managed" flag
to keep track of whether or not it's on csched's alist. As a result, sp3_node_unlink_all()
is now idempotent as there is at least one place (from prune trees) where it could be
called many times for the same node. We can also call it to reap a node from a join, and
I added some code to check and ensure that node isn't also on the active dirty list and
to deal with it if it is (we don't expect it to be on that list).

Signed-off-by: Greg Becker <gbecker@micron.com>

Signed-off-by: Greg Becker <gbecker@micron.com>