Add support for stashing the CLI program names in a common way
Many of our tools store the name of the program, typically for use in
usage statements. All of the programs that store the name did so in
custom ways and created repeated code.

The pattern we use fairly often, but not always, in CLIs is output
taking the form of `<progname>: <message>`. By storing the program name
in libhse-cli, we can more effectively move to a paradigm where we don't
create multiple `fatal()`, `syntax()`, etc. functions per program,
though this improvement can come at a later time.

Signed-off-by: Tristan Partin <tpartin@micron.com>