Codify some hsettp heurisitcs that Greg added
Greg changed the exact match search for an operation to a best match at
some point. I think the idea is pretty solid, but a casual user would
never know about it. A casual user is anyone but Greg for the record :).

Another thing that Greg added was hiding certain operations from the
general help output because their names were too long. For instance,
getting perf counter events in a certain set in a certain file on a
certain line. What a mouthful! So I went ahead and hid obvious
candidates instead of relying on the length heuristic that Greg
originally went by.

Signed-off-by: Tristan Partin <tpartin@micron.com>