Remove defining _LGPL_SOURCE
To comply with a statically linked LGPL library like userspace-rcu, one
needs to provide the ability to build and re-link hse. Previously, we
were defining _LGPL_SOURCE all the time, but people potentially using
and distributing HSE might not have known about this. Turning it off by
is a safer alternative.

Signed-off-by: Tristan Partin <tpartin@micron.com>